1.1 Дипломный проект BLOG-ENGINE v0.14 GET	/api/post - переделан под Hibernate + pageable/sort GET	/api/settings - переделан под Hibernate + начальные значения вынесены в properties GET	/api/post/search - переделан под Hibernate + pageable/sort GET	/api/post/byDate - переделан под Hibernate + pageable/sort GET	/api/post/byTag	- переделан под Hibernate + pageable/sort + развернут FlyWay со старт.скриптами

+++ Blog_Engine/src/main/java/org/skillbox/springbootrest/service/UserService.java
@@ -53,7 +53,7 @@ public class UserService {
             return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
         }
         User user = new User();
-        user.setIsModerator(0);
+        user.setIsModerator((byte) 0);
         user.setRegTime(Timestamp.from(Instant.now()));
         user.setName(registerRequest.getName());
         user.setEmail(registerRequest.geteMail());
